feat: add settings sheet for AI provider configuration
