--- core/src/zap/ai/components/model-select.tsx
@@ -1,60 +1,0 @@
-import type { Control } from "react-hook-form";
-
-import {
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/components/ui/form";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
-import { ModelsByProvider } from "@/zap/data/ai";
-import type { AIFormValues, AIProviderId } from "@/zap/types/ai.types";
-
-interface ModelSelectProps {
-  control: Control<AIFormValues>;
-  disabled: boolean;
-  provider: AIProviderId;
-}
-
-export function ModelSelect({ control, disabled, provider }: ModelSelectProps) {
-  const models = ModelsByProvider[provider];
-
-  return (
-    <FormField
-      control={control}
-      name="model"
-      render={({ field }) => (
-        <FormItem>
-          <FormLabel>Model</FormLabel>
-          <Select
-            disabled={disabled}
-            onValueChange={field.onChange}
-            value={field.value}
-          >
-            <FormControl>
-              <SelectTrigger>
-                <SelectValue placeholder="Select a model" />
-              </SelectTrigger>
-            </FormControl>
-
-            <SelectContent>
-              {models.map((model) => (
-                <SelectItem key={model} value={model}>
-                  {model}
-                </SelectItem>
-              ))}
-            </SelectContent>
-          </Select>
-          <FormMessage />
-        </FormItem>
-      )}
-    />
-  );
-}
