--- core/src/zap/ai/components/provider-select.tsx
@@ -1,61 +1,0 @@
-import type { Control } from "react-hook-form";
-
-import {
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/components/ui/form";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
-import { AI_PROVIDERS_OBJECT } from "@/zap/data/ai";
-import { AIProviderIdSchema } from "@/zap/schemas/ai.schema";
-import type { AIFormValues } from "@/zap/types/ai.types";
-
-export interface FormFieldProps {
-  control: Control<AIFormValues>;
-  disabled: boolean;
-}
-
-export function ProviderSelect({ control, disabled }: FormFieldProps) {
-  return (
-    <FormField
-      control={control}
-      name="provider"
-      render={({ field }) => (
-        <FormItem>
-          <FormLabel>AI Provider</FormLabel>
-          <Select
-            disabled={disabled}
-            onValueChange={field.onChange}
-            value={field.value}
-          >
-            <FormControl>
-              <SelectTrigger>
-                <SelectValue placeholder="Select a provider" />
-              </SelectTrigger>
-            </FormControl>
-
-            <SelectContent>
-              {AIProviderIdSchema.options.map((provider) => (
-                <SelectItem key={provider} value={provider}>
-                  {
-                    AI_PROVIDERS_OBJECT.find((p) => p.provider === provider)
-                      ?.name
-                  }
-                </SelectItem>
-              ))}
-            </SelectContent>
-          </Select>
-          <FormMessage />
-        </FormItem>
-      )}
-    />
-  );
-}
